Move last vehicle db stuff from Future to DBIO.
